refactor(blocks-construt): revamp the calendar
