+++ .gitignore
@@ -24,6 +24,7 @@ yarn-error.log*
 .env.local
 
 # IDE - VSCode
+.cursor/*
 .vscode/*
 !.vscode/settings.json
 !.vscode/tasks.json
