+++ package.json
@@ -24,6 +24,7 @@
     "@hcaptcha/react-hcaptcha": "^1.11.2",
     "@hookform/resolvers": "^3.10.0",
     "@radix-ui/react-alert-dialog": "^1.1.5",
+    "@radix-ui/react-avatar": "^1.1.3",
     "@radix-ui/react-checkbox": "^1.1.2",
     "@radix-ui/react-collapsible": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.4",
@@ -31,6 +32,7 @@
     "@radix-ui/react-dropdown-menu": "^2.1.2",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.1.1",
+    "@radix-ui/react-menubar": "^1.1.6",
     "@radix-ui/react-popover": "^1.1.2",
     "@radix-ui/react-radio-group": "^1.2.0",
     "@radix-ui/react-scroll-area": "^1.1.0",
@@ -41,8 +43,6 @@
     "@radix-ui/react-tabs": "^1.1.0",
     "@radix-ui/react-toast": "^1.2.2",
     "@radix-ui/react-tooltip": "^1.1.3",
-    "@radix-ui/react-avatar": "^1.1.3",
-    "@radix-ui/react-menubar": "^1.1.6",
     "@tanstack/react-query": "^5.66.0",
     "@tanstack/react-query-devtools": "^5.59.16",
     "@tanstack/react-table": "^8.20.6",
@@ -70,6 +70,7 @@
     "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
     "recharts": "^2.15.0",
+    "rrule": "^2.8.1",
     "tailwind-merge": "^2.6.0",
     "zod": "^3.24.1",
     "zustand": "^5.0.2"
