+++ src/features/big-calendar/components/big-calendar-header/big-calendar-header.tsx
@@ -1,0 +1,98 @@
+import { useEffect, useState } from 'react';
+import { Search, ListFilter, Plus } from 'lucide-react';
+import { SlotInfo } from 'react-big-calendar';
+import { DateRange } from 'react-day-picker';
+import { Button } from 'components/ui/button';
+import { Dialog } from 'components/ui/dialog';
+import { Input } from 'components/ui/input';
+import { AddEvent } from '../modals/add-event/add-event';
+import { CalendarFilterSheet } from '../calendar-filters-sheet/calendar-filters-sheet';
+
+interface BigCalendarHeaderProps {
+  title?: string;
+  onAddEvent: () => void;
+  selectedSlot: SlotInfo | null;
+  onEventSubmit: (data: { title: string; start: string; end: string }) => void;
+  onDialogClose: () => void;
+  onApplyFilters: (filters: { dateRange: DateRange; color: string | null }) => void;
+  searchPlaceholder?: string;
+  onSearchChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
+
+export const BigCalendarHeader = ({
+  title = 'Calendar',
+  onAddEvent,
+  selectedSlot,
+  onEventSubmit,
+  onDialogClose,
+  searchPlaceholder = 'Search',
+  onSearchChange,
+  onApplyFilters,
+}: Readonly<BigCalendarHeaderProps>) => {
+  const [openSheet, setOpenSheet] = useState(false);
+
+  const handleFilters = () => {
+    setOpenSheet(true);
+  };
+
+  useEffect(() => {
+    if (openSheet) {
+      document.body.style.overflow = 'hidden';
+    } else {
+      document.body.style.overflow = '';
+    }
+
+    return () => {
+      document.body.style.overflow = '';
+    };
+  }, [openSheet]);
+
+  return (
+    <>
+      <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-2">
+        <h1 className="text-2xl font-bold leading-9">{title}</h1>
+        <div className="flex items-center gap-2 sm:justify-end">
+          <div className="relative w-[45%]">
+            <Search className="absolute left-3 top-1/2 h-3 w-3 -translate-y-1/2 bg-background" />
+            <Input
+              placeholder={searchPlaceholder}
+              className="h-8 w-full rounded-lg bg-background pl-8"
+              onChange={onSearchChange}
+            />
+          </div>
+          <Button
+            variant="outline"
+            size="sm"
+            className="text-sm font-bold sm:min-w-[116px]"
+            onClick={handleFilters}
+          >
+            <ListFilter className="w-5 h-5" />
+            <span className="sr-only sm:not-sr-only">Filters</span>
+          </Button>
+          <Button size="sm" onClick={onAddEvent} className="text-sm font-bold sm:min-w-[116px]">
+            <Plus className="w-5 h-5" />
+            <span className="sr-only sm:not-sr-only">Add Event</span>
+          </Button>
+        </div>
+      </div>
+      <Dialog open={!!selectedSlot} onOpenChange={onDialogClose}>
+        {selectedSlot && (
+          <AddEvent
+            start={selectedSlot.start}
+            end={selectedSlot.end}
+            onSubmit={onEventSubmit}
+            onCancel={onDialogClose}
+          />
+        )}
+      </Dialog>
+      <CalendarFilterSheet
+        open={openSheet}
+        onOpenChange={setOpenSheet}
+        onApplyFilters={(filters) => {
+          onApplyFilters?.(filters);
+          setOpenSheet(false);
+        }}
+      />
+    </>
+  );
+};
