+++ src/features/big-calendar/constants/calendar.constants.ts
@@ -1,0 +1,9 @@
+export type CustomView = 'month' | 'week' | 'day' | 'agenda' | 'year';
+
+export const CALENDAR_VIEWS: CustomView[] = ['agenda', 'day', 'week', 'month', 'year'];
+
+export const WEEK_DAYS = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
+
+export const WEEK_DAYS_RRULE = ['SU', 'MO', 'TU', 'WE', 'TH', 'FR', 'SA'];
+
+export const CALENDER_PERIOD = ['Day', 'Week', 'Month', 'Year'];
