+++ src/features/big-calendar/enums/calendar.enum.ts
@@ -1,0 +1,36 @@
+export enum CalendarEventColor {
+  PRIMARY = 'hsl(var(--primary-500))',
+  SECONDARY = 'hsl(var(--secondary-500))',
+  DEEPPURPLE = 'hsl(var(--purple-500))',
+  BURGUNDY = 'hsl(var(--burgundy-500))',
+  WARNING = 'hsl(var(--warning))',
+  PRIMARY100 = 'hsl(var(--primary-100))',
+  SECONDARY100 = 'hsl(var(--secondary-100))',
+  DEEPPURPLE100 = 'hsl(var(--purple-100))',
+  BURGUNDY100 = 'hsl(var(--burgundy-100))',
+}
+
+export enum MEMBER_STATUS {
+  ACCEPTED = 'accepted',
+  DECLINED = 'declined',
+  NORESPONSE = 'no response',
+}
+
+export enum CalendarModalState {
+  NONE,
+  EVENT_DETAIL,
+  EDIT_EVENT,
+  EVENT_RECURRENCE,
+}
+
+export enum EventContentTextColor {
+  PRIMARY = 'hsl(var(--white))',
+  SECONDARY = 'hsl(var(--high-emphasis))',
+  DEEPPURPLE = 'hsl(var(--white))',
+  BURGUNDY = 'hsl(var(--white))',
+  WARNING = 'hsl(var(--high-emphasis))',
+  PRIMARY100 = 'hsl(var(--high-emphasis))',
+  SECONDARY100 = 'hsl(var(--high-emphasis))',
+  DEEPPURPLE100 = 'hsl(var(--high-emphasis))',
+  BURGUNDY100 = 'hsl(var(--high-emphasis))',
+}
