+++ src/features/big-calendar/types/calendar-event.types.ts
@@ -1,0 +1,23 @@
+import { Event } from 'react-big-calendar';
+import { MEMBER_STATUS } from '../enums/calendar.enum';
+
+export interface Member {
+  id: string;
+  name: string;
+  image: string;
+  status?: MEMBER_STATUS;
+}
+export interface CalendarEvent extends Event {
+  eventId?: string;
+  title: string;
+  start: Date;
+  end: Date;
+  allDay?: boolean;
+  resource?: {
+    meetingLink?: string;
+    description?: string;
+    color?: string;
+    members?: Member[];
+    recurring?: boolean;
+  };
+}
