+++ src/features/big-calendar/utils/locales.ts
@@ -1,0 +1,19 @@
+import { dateFnsLocalizer } from 'react-big-calendar';
+import { format, getDay, parse, startOfWeek } from 'date-fns';
+import enUS from 'date-fns/locale/en-US';
+
+const locales = {
+  'en-US': enUS,
+};
+
+export const calendarLocalizer = dateFnsLocalizer({
+  format,
+  parse,
+  startOfWeek,
+  getDay,
+  locales,
+});
+
+export const calendarTimeFormat = {
+  timeGutterFormat: 'HH:mm',
+};
