--- src/features/calendar/components/big-calendar/big-calendar.tsx
@@ -1,5 +1,0 @@
-import { Calendar } from 'react-big-calendar';
-import 'react-big-calendar/lib/addons/dragAndDrop/styles.css';
-import './big-calendar.css';
-
-export const BigCalendar = Calendar;
