--- src/features/calendar/constants/calendar.constants.ts
@@ -1,7 +1,0 @@
-export type CustomView = 'month' | 'week' | 'day' | 'agenda' | 'year';
-
-export const CALENDAR_VIEWS: CustomView[] = ['agenda', 'day', 'week', 'month', 'year'];
-
-export const WEEK_DAYS = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
-
-export const CALENDER_PERIOD = ['Day', 'Week', 'Month', 'Year'];
