--- src/features/calendar/enums/calendar.enum.ts
@@ -1,36 +1,0 @@
-export enum CalendarEventColor {
-  PRIMARY = 'primary',
-  SECONDARY = 'secondary',
-  DEEPPURPLE = 'purple',
-  BURGUNDY = 'burgundy',
-  WARNING = 'warning',
-  PRIMARY100 = 'primary-100',
-  SECONDARY100 = 'secondary-100',
-  DEEPPURPLE100 = 'purple-100',
-  BURGUNDY100 = 'burgundy-100',
-}
-
-export enum MEMBER_STATUS {
-  ACCEPTED = 'accepted',
-  DECLINED = 'declined',
-  NORESPONSE = 'no response',
-}
-
-export enum CalendarModalState {
-  NONE,
-  EVENT_DETAIL,
-  EDIT_EVENT,
-  EVENT_RECURRENCE,
-}
-
-export enum EventContentTextColor {
-  PRIMARY = 'text-white',
-  SECONDARY = 'text-high-emphasis',
-  DEEPPURPLE = 'text-white',
-  BURGUNDY = 'text-white',
-  WARNING = 'text-high-emphasis',
-  PRIMARY100 = 'text-high-emphasis',
-  SECONDARY100 = 'text-high-emphasis',
-  DEEPPURPLE100 = 'text-high-emphasis',
-  BURGUNDY100 = 'text-high-emphasis',
-}
