--- src/features/calendar/index.ts
@@ -1,18 +1,0 @@
-export * from './utils/locales';
-export * from './utils/date-utils';
-export * from './enums/calendar.enum';
-export * from './services/calendar-services';
-export * from './types/calendar-event.types';
-export * from './constants/calendar.constants';
-export { BigCalendar } from './components/big-calendar/big-calendar';
-export { AddEvent } from './components/modals/add-event/add-event';
-export { BigCalendarHeader } from './components/big-calendar-header/big-calendar-header';
-export { CalendarToolbar } from './components/calendar-toolbar/calendar-toolbar';
-export { CalendarFilterSheet } from './components/calendar-filters-sheet/calendar-filters-sheet';
-export { ColorPickerTool } from './components/color-picker-tool/color-picker-tool';
-export { EventDetails } from './components/modals/event-details/event-details';
-export { EditRecurrence } from './components/modals/edit-recurrence/edit-recurrence';
-export { EventsContent } from './components/events-content/events-content';
-export { EditEvent } from './components/modals/edit-event/edit-event';
-export { AgendaContent } from './components/agenda-content/agenda-content';
-export { YearContent } from './components/year-content/year-content';
