--- src/features/calendar/utils/locales.ts
@@ -1,15 +1,0 @@
-import { dateFnsLocalizer } from 'react-big-calendar';
-import { format, getDay, parse, startOfWeek } from 'date-fns';
-import enUS from 'date-fns/locale/en-US';
-
-const locales = {
-  'en-US': enUS,
-};
-
-export const localizer = dateFnsLocalizer({
-  format,
-  parse,
-  startOfWeek,
-  getDay,
-  locales,
-});
